# java-server/src/main/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: Optional multiline or single-line description in [CommonMark](http://commonmark.org/help/)
    or HTML.
  title: Sample API
  version: 0.1.9
servers:
- description: Optional server description, e.g. Main (production) server
  url: http://localhost:8080/myapp/
- description: Optional server description, e.g. Internal staging server for testing
  url: http://staging-api.example.com
paths:
...
  /files/download/{fileId}:
    get:
      operationId: download
      parameters:
      - description: The ID of the download file to return.
        explode: false
        in: path
        name: fileId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "500":
          content:
            application/json:
              schema:
                type: string
          description: Error
      x-accepts: application/json
...